strategy_bot/requirements.txt
pandas-ta
aiogram
scikit-learn
yfinance
pytz
matplotlib
ccxt
asyncpg
torch